services/lip_sync/requirements.ls.txt
ffmpeg==1.4
librosa==0.10.1
scikit-image==0.22.0
facenet-pytorch==2.5.3
onnxruntime-gpu==1.17.1
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
